+++ CHANGES.rst
@@ -11,6 +11,15 @@ Changelog of threedi-wms
 0.9 (2014-03-06)
 ----------------
 
+- Implement volume mask and alpha channel for masked arrays.
+
+- Improved getmap messages.
+
+- Getprofile, gettimeseries now also works with messages. Gettimeseries uses
+  messages for getting the height and quad cell.
+
+- Removed gettimeseries:timestep option.
+
 - Improved memory messages, ports are configurable in gunicorn: bin/gunicorn 'server.app:build_app(req_port=5557,sub_port=5558)' -w 1 -b 0.0.0.0:5000